Service implemented
Got service up and running.  Currently only receives and passes back
the URL for our API (as a test to make sure it works), so next we need
to have it actually pull down data with the url.  Also, redid the
earlier handler to utilize an inner class to avoid memory leaks as per
the project rubric requirements.