Now have remote data
Implemented necessary logic within the service class to retrieve the
remote data.  It’s currently a string, which I will send to the
FileManager singleton class (Not written yet), and since we now have a
way to alert our MainActivity when the data is done (and saved!), we
can then pull it in for display.  Woot!