File Manager done
File Manager is completed.  Successfully saves raw JSON data in string
format to file on device storage, and has the read method implemented
as well.  Next will be calling the read method from the handler
function in MainActivity (after confirming the data pull was successful
in the service), and converting to JSON/getting data we can use.