Formatted JSON data for display
Formatted JSON data that is read from the FileManager class to get the
data we want for display.  Each sports story has it’s own hash map to
hold all data relevant to it, which are then contained within an array
of hash maps, as demonstrated by the tutorial video.  Next, it’s time
to finally build out the interface to display this data to the user.