Added reload button and Saved data check
Added a reload button to dynamically display in the event the user has
no internet connection AND no saved data.  This button basically
rechecks internet status and then starts the retrieval of data is
internet is found.  Also created a check to determine if a file exists
at the path the FileManager singleton class is looking for it at, that
way we don’t have any crashes/runtime errors.