Rating alert dialog done
Finished creating a basic alert dialog to display when the DetailView
is dismissed.  It currently displays the title of the previously viewed
story, and uses an non-editable rating bar to display what was
previously rated.  Very little styling was done.  Will get to it in the
future.  Next up is the implicit intent.