Save state on DetailsView
Finished the saved state functionality for the DetailsView so that we
can save the image that was pulled down from the internet, instead of
re-downloading it over and over.  Next will create a back button to
handle exiting the DetailView activity, even though the back button
currently works as well.