Alert Dialog Saves on rotation
FINALLY figured out a way to have alert dialog save its data and
redisplay to the user upon rotating the device.  There was a strange
bug where the data was being erased after rotating the device twice.
Implemented logic to manually reassign the values before the view is
destroyed, and made sure to manually destroy the alert dialog when the
user taps the “Okay” button.  Whew.  Next will create a back button for
a user to leave the DetailsView.  Then code cleanup/comments.