Implemented ActionBar back button
Added a back button to the DetailsView activity within the ActionBar to
make it look a bit more standard.  The user can now return to
MainActivity using this button, or the standard Android back button.
Both call the same finish() method and pass the intent/data back to the
MainActivity.