Removed all UI from MainActivity
Removed all instances of UI elements from the MainActivity class and
migrated them over to the MainActivity fragment.  Created methods for
the MainActivity to update the interface using the interface methods
provided by the MainActivityFragment.  The only bit of UI still
contained within MainActivity is the Alert Dialog, which will probably
be migrated over to a fragment last.