Fixed error in saving state of DetailsFragment
Fixed an error where if the user rotates the device while the
DetailsFragment is being displayed (fullscreen), the title and rating
aren’t saved in time and subsequently not passed back through the
intent in finish().  This resulted in blank alert dialogs.  Need to fix
an error where this same scenario CAN result in previous alert dialogs
being displayed as well.