Fixed multiple alert dialogs
Fixed an uncommon bug where multiple alert dialogs could show with the
new landscape layout.  Need to decide whether or not to move the alert
dialog logic over to the MainFragment or not, since it’s not really
related to that interface at all (nor is its data).  Also want to
create a custom styling for the DetailFragment in the event the user
hasn’t selected a story from the MainActivityFragment’s listView yet.