Created search functionality
Create a custom DialogFragment to handle the search functionality.  It
accepts a string and then using the MainActivity, passes the string to
the MainActivity Fragment, which we then manually use to filter our
list view.  Current implementation is ugly and desperately needs to be
refactored.  Next I want to create a footer for the list view that
displays the number of results, and lets the user reset the list view
to the default stories.