Created listView footer/finalized search
Created a listView footer to display data to the user.  If no search is
done (default state), it simply shows how many stories were retrieved.
If a search is done, it shows the number of results, and allows the
user to reset the list to its default state.  Next I need to fix the
onItemClick method of the listView to pass the correct story to
MainActivity for display within the DetailsView Activity/Fragment.