Preferences Done
Created dialog fragment for the “preferences” requirement.  Allow user
to enter their name and favorite sport.  User’s name is then displayed
next to a welcome message above the list view.  Preferences fragment
has been coded to recognize whether or not data already exits, and if
so, alter it’s interface to allow the user to change the data later,
without it appearing as if the app is running again for the first time.