Favorites activity and saving to file
Created an activity to display the user’s saved favorites, which are
grabbed from the detailView and saved out to a file.  Selecting the
‘favorite’ icon from the MainActivity in portrait by default and a
choice in landscape loads the Favorites activity, which then calls
FileManager singleton again to load in the saved stories into a
listView.  Would like to reuse the DetailsActivity if there’s time for
the user to completely load the stories they’ve saved from the
Favorites activity as if they were new stories.