Favorites Activity list view done
Finalized the code to display favorited/saved stories within a list
view in the Favorites activity.  Selecting the stories does nothing,
but if I have time, will hook up to “reuse” the DetailsView so the user
can view the additional details for each story as well.  Also completed
adding the favorites icon to the Action Bar, and all code required for
the user to delete their favorites from this Activity.  ListView
updates in realtime as items are removed.