Finalized app code
Created info activity to display basic info about the app.  Styled it
to look like a popover, almost like an alert.  Redid the list view
within the “Favorites” activity to display the rating the user saved
within each row and added image assets to project to support this.
Hooked up the Details Activity to allow displaying of “favorited” story
data as well as loading in the data from the MainActivity’s list view
as well.  App should be functionally complete.  Need to add in a
liberal amount of comments, and then turn in.