Fixed bug I've been looking for forever
Fixed a bug that would crash the app if the user selected either the
header or footer of the MainActivity’s list view.  I had no idea that
the onClickMethod could detect these items as being tappable.  This
resulted in an incorrect value being sent to MainActivity, which caused
an “OutOfBounds” error on an array it was referenced against, crashing
the app.